Introduce get_client_ip
