+++ saleor/core/utils/__init__.py
@@ -1,11 +1,15 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
+from babel.numbers import get_territory_currencies
 from django import forms
 from django.conf import settings
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.utils.encoding import iri_to_uri, smart_text
+from django_countries import countries
+from django_countries.fields import Country
 from django_prices_openexchangerates import exchange_currency
+from geolite2 import geolite2
 from prices import PriceRange
 
 try:
@@ -36,6 +40,30 @@ def build_absolute_uri(location, is_secure=False):
     current_uri = '%s://%s' % ('https' if is_secure else 'http', host)
     location = urljoin(current_uri, location)
     return iri_to_uri(location)
+
+
+def get_client_ip(request):
+    ip = request.META.get('HTTP_X_FORWARDED_FOR', None)
+    if ip:
+        return ip.split(',')[0].strip()
+    return request.META.get('REMOTE_ADDR', None)
+
+
+def get_country_by_ip(ip_address):
+    reader = geolite2.reader()
+    geo_data = reader.get(ip_address)
+    geolite2.close()
+    if geo_data and 'country' in geo_data and 'iso_code' in geo_data['country']:
+        country_iso_code = geo_data['country']['iso_code']
+        if country_iso_code in countries:
+            return Country(country_iso_code)
+
+
+def get_currency_for_country(country):
+    currencies = get_territory_currencies(country.code)
+    if len(currencies):
+        return currencies[0]
+    return settings.DEFAULT_CURRENCY
 
 
 def get_paginator_items(items, paginate_by, page):
